Paging support in TileDB See variant/example/src/example_libtiledb_variant_driver.cc for an example of how to use paging. Variant paging depends on a few assumptions 1. Cells are sorted in column major order and cells in the same column are sorted by row idx. 2. Variants are returned in order by gt_get_column_interval() function 3. Calls that are merged into a Variant have the same start position.
The GA4GHPagingInfo structure stores the page size (#variants to return
per call - note #variants NOT #VariantCalls). Whenever this limit is
hit, the co-ordinates of the last cell accessed are stored. The next
call will access cells that are after the stored co-ordinate ("after"
refers to the ordering described above).

Some cells are re-iterated in subsequent calls in the left-sweep
operation and the forward sweep also. However, these cells will not be
returned in the subsequent call.