When using the Java interface for loading VCF data as a stream, callsets information need not be specified in the callsets JSON. Instead, the information can be passed as a Map<int, SampleInfo> where the key corresponds to the index of the sample in the VCF header and SampleInfo contains the TileDB row index and a globally unique name for the sample (unique across all streams/files).
The above information is serialized to a string in JSON format, passed
to the C++ layers where it is read (same function that parses the
callsets JSON file). All intermediate classes now contain an additional
string argument to pass this information.

Use simplejson Java library for JSON handling - added dependency in
pom.xml.