Goal: Maintain columnar structure as much as possible while still supporting different scanning patterns (VCF like output)
While TileDB supports scanning by cells, GenomicsDB requires scans of
different types to be supported. The main issue is that a cell may be
'live' in GenomicsDB even after succeeding cells are processed. The
reason being that 'cells' in GenomicsDB are intervals and not single
co-ordinate cells.

To maintain columnar structure, we define the GenomicsDBColumnarField class
to contain not one buffer, but two linked lists of GenomicsDBBuffers - free
and live. The live list holds all buffers which have at least one live
cell - the list itself is maintained in column major order. The free
list holds allocated buffers which do not contain live data.

The GenomicsDBColumnarField class contains no intelligence about how to
maintain the live and free lists. That functionality is in the iterator
classes. Since each iterator has a different scan pattern, it must be
responsible for correct maintenance of the lists (and the indexes in
each live buffer).

* Incomplete implementation of GenomicsDBBuffers
* Incomplete implementation of GenomicsDBColumnarField
* Incomplete implementation of SingleCellIterator.