To deal with the sweep operation, added a structure called GenomicsDBLiveCellMarker. For each row in the array, this structure stores the cell that intersects the query column interval begin. To keep the columnar structure, GenomicsDBLiveCellMarker only stores pointers to the GenomicsDBBuffer objects (and indexes). By using the live cell counter in GenomicsDBBuffer, live buffers are tracked and moved to the free list only when the live counter == 0.
The SingleCellTileDBIterator runs in 2 modes:
a. Normal forward iteration with bounds on dimension
b. Find intersecting intervals mode - this is invoked when a query
column interval is provided. In this mode, cells intersecting with the
query interval begin are determined. Such cells are tracked using the
GenomicsDBLiveCellMarker structure and are organized into a PQ
(column-major order).