Corner case bug fix - hat tip to Ramesh (@rsm0001) for exercising the corner case.
If the query_attributes field in the query JSON does not contain REF and
ALT fields, but contains a field like PL which requires REF and ALT,
then these fields are added to the list of query fields. However, the
bug is that the code doesn't specify that these fields (REF, ALT) are variable
length fields when added.

Note that if the REF and ALT fields were part of the initial queried list,
they would have been marked as variable length fields. Hence, the bug
wasn't seen in the common cases.