Consolidation of GenomicsDB configurations - simplified class hierarchy for configuration objects.
GenomicsDBConfigBase - base class holding practically all the
information needed for queries and imports

GenomicsDBImportConfig - subclass of GenomicsDBConfigBase that contains
some import specific information

VariantQueryConfig - subclass of GenomicsDBConfigBase

This way all the information is consolidated into a couple of classes.

Most of the changes in this commit are to handle the updated class
hierarchy and to avoid duplication of config information as much as
possible.

TODO: Java side changes